start structuring nym js lib
