+++ packages/nymjs/README.md
@@ -1,0 +1,5 @@
+js lib for creating/interacting with nyms
+
+modeled after https://github.com/personaelabs/spartan-ecdsa/tree/main/packages/lib
+
+TODO: actual docs on lib usage
