+++ packages/nymjs/src/lib.ts
@@ -1,0 +1,2 @@
+export * from "./types";
+export * from "./prover";
