+++ packages/nymjs/src/prover.ts
@@ -1,42 +1,11 @@
 // NOTE: this file should be a reference for frontend code
 import { hashPersonalMessage, fromRpcSig } from "@ethereumjs/util";
 
-// NOTE: wagmi signMessage uses Wallet.signMessage under the hood
-import { Wallet } from "ethers";
-
 import { computeEffEcdsaPubInput } from "@personaelabs/spartan-ecdsa";
 
 import { buildPoseidon } from "circomlibjs";
 
-// NOTE: OBVIOUSLY DON'T USE THIS ANYWHERE! IT'S COMPROMISED.AF
-export const testPrivateKey =
-  "0x58d23b55bc9cdce1f18c2500f40ff4ab7245df9a89505e9b1fa4851f623d241d";
-export const testSigner = new Wallet(testPrivateKey);
-
-export type EffECDSASig = {
-  Tx: bigint;
-  Ty: bigint;
-  Ux: bigint;
-  Uy: bigint;
-  s: bigint; // private
-};
-
-export type MerkleProof = {
-  root: bigint;
-  siblings: bigint[];
-  pathIndices: number[];
-};
-
-export type NymInput = {
-  membershipProof: MerkleProof;
-
-  nym: string;
-  nymHash: string;
-  nymSig: EffECDSASig;
-
-  content: string;
-  contentSig: EffECDSASig;
-};
+import { MerkleProof, NymInput, EffECDSASig } from "./types";
 
 function bufferToBigInt(buf: Buffer): bigint {
   return BigInt("0x" + buf.toString("hex"));
@@ -59,12 +28,11 @@ async function computeEffECDSASig(
 }
 
 async function computeNymHash(nymSigS: bigint) {
-  // TODO: re-use
+  // TODO: re-use Poseidon instance
   const poseidon = await buildPoseidon();
   return poseidon([nymSigS, nymSigS]);
 }
 
-// NOTE: this method may not need be exposed, just generate proof
 export async function computeNymInput(
   membershipProof: MerkleProof,
 
@@ -92,4 +60,11 @@ export async function computeNymInput(
   };
 }
 
-// TODO: generate proof method
+// NOTE: better name?
+// TODO: return proof type
+export async function proveNym(input: NymInput) {
+  // TODO: generate witness using typed input (requires witness gen wasm)
+  // TODO: load circuit from google storage
+  // TODO: get public input in bytes
+  // TODO: do proof with spartan wasm (requires spartan wasm bytes to be in repo)
+}
